# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PetConnect API
  description: |
    API REST para PetConnect — plataforma de mascotas que conecta dueños,
    veterinarios y comunidad. Incluye feed social, ranking semanal, InstaPet
    (red social por mascota), mascotas perdidas y adopciones.
  version: 0.1.0
  contact:
    name: PetConnect Team
servers:
  - url: http://localhost:8001/api/v1
    description: Servidor local de desarrollo (Hono)
  - url: https://api.petconnect.app/api/v1
    description: Producción
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar nuevo usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [Auth]
      summary: Obtener perfil del usuario autenticado
      responses:
        "200":
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      tags: [Auth]
      summary: Actualizar perfil propio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Perfil actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # =========================================================================
  # PETS
  # =========================================================================
...
  /lost-pets:
    get:
      tags: [LostPets]
      summary: Listar mascotas perdidas
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            type: string
            enum: [lost, found]
          description: Filtrar por estado
      responses:
        "200":
          description: Lista paginada de mascotas perdidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLostPets"
    post:
      tags: [LostPets]
      summary: Reportar mascota perdida
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLostPetRequest"
      responses:
        "201":
          description: Reporte creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostPet"
        "422":
          $ref: "#/components/responses/ValidationError"
  /lost-pets/{id}:
    get:
      tags: [LostPets]
      summary: Obtener detalle de mascota perdida
      parameters:
        - $ref: "#/components/parameters/LostPetId"
      responses:
        "200":
          description: Detalle de mascota perdida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostPetDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [LostPets]
      summary: Actualizar reporte de mascota perdida
      parameters:
        - $ref: "#/components/parameters/LostPetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLostPetRequest"
      responses:
        "200":
          description: Reporte actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostPet"
    delete:
      tags: [LostPets]
      summary: Eliminar reporte
      parameters:
        - $ref: "#/components/parameters/LostPetId"
      responses:
        "204":
          description: Reporte eliminado

  # =========================================================================
  # ADOPTIONS
  # =========================================================================
  /adoptions:
    get:
      tags: [Adoptions]
      summary: Listar adopciones
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            type: string
            enum: [available, pending, adopted]
          description: Filtrar por estado
      responses:
        "200":
          description: Lista paginada de adopciones
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAdoptions"
    post:
      tags: [Adoptions]
      summary: Publicar mascota en adopción
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdoptionRequest"
      responses:
        "201":
          description: Adopción publicada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adoption"
        "422":
          $ref: "#/components/responses/ValidationError"
  /adoptions/{id}:
    get:
      tags: [Adoptions]
      summary: Obtener detalle de adopción
      parameters:
        - $ref: "#/components/parameters/AdoptionId"
      responses:
        "200":
          description: Detalle de adopción
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdoptionDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Adoptions]
      summary: Actualizar publicación de adopción
      parameters:
        - $ref: "#/components/parameters/AdoptionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdoptionRequest"
      responses:
        "200":
          description: Adopción actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adoption"
    delete:
      tags: [Adoptions]
      summary: Eliminar publicación de adopción
      parameters:
        - $ref: "#/components/parameters/AdoptionId"
      responses:
        "204":
          description: Publicación eliminada

  # =========================================================================
  # INSTAPET SOCIAL
  # =========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT de Supabase Auth

  # ------------------------------------------------------------------
  # Parameters (reusables)
  # ------------------------------------------------------------------
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Número de página
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
      description: Resultados por página
...
    LostPetId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AdoptionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Conflicto (registro duplicado)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ------------------------------------------------------------------
  # Schemas
  # ------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Recurso no encontrado"
      required: [detail]
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        pages:
          type: integer
          example: 3

    # --- Auth ---
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "usuario@email.com"
        password:
          type: string
          minLength: 6
          example: "SecurePass123"
        username:
          type: string
          example: "petlover"
      required: [email, password, username]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "usuario@email.com"
        password:
          type: string
          example: "SecurePass123"
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIs..."
        token_type:
          type: string
          example: "bearer"
        profile:
          $ref: "#/components/schemas/Profile"
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_id:
          type: string
          format: uuid
        username:
          type: string
          example: "petlover"
        avatar_url:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Pet ---
    Pet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: string
          example: "Luna"
        species:
          type: string
          example: "Perro"
        breed:
          type: string
          nullable: true
          example: "Golden Retriever"
        age:
          type: integer
          nullable: true
          example: 24
        weight:
          type: number
          nullable: true
          example: 25.5
        photo_url:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    LostPet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reporter_id:
          type: string
          format: uuid
        name:
          type: string
          example: "Max"
        species:
          type: string
          example: "Gato"
        breed:
          type: string
          nullable: true
        photo_url:
          type: string
          format: uri
          nullable: true
        last_seen_lat:
          type: number
          format: double
          example: 19.4326018
        last_seen_lng:
          type: number
          format: double
          example: -99.1332093
        last_seen_address:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [lost, found]
          example: "lost"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LostPetDetail:
      allOf:
        - $ref: "#/components/schemas/LostPet"
        - type: object
          properties:
            reporter:
              $ref: "#/components/schemas/Profile"
    CreateLostPetRequest:
      type: object
      properties:
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        photo_url:
          type: string
          format: uri
        last_seen_lat:
          type: number
          format: double
        last_seen_lng:
          type: number
          format: double
        last_seen_address:
          type: string
        description:
          type: string
      required: [name, species, last_seen_lat, last_seen_lng]
    UpdateLostPetRequest:
      type: object
      properties:
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        photo_url:
          type: string
          format: uri
        last_seen_lat:
          type: number
          format: double
        last_seen_lng:
          type: number
          format: double
        last_seen_address:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [lost, found]
    PaginatedLostPets:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/LostPet"

    # --- Adoption ---
    Adoption:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pet_id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        adopter_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [available, pending, adopted]
          example: "available"
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AdoptionDetail:
      allOf:
        - $ref: "#/components/schemas/Adoption"
        - type: object
          properties:
            pet:
              $ref: "#/components/schemas/Pet"
            owner:
              $ref: "#/components/schemas/Profile"
            adopter:
              $ref: "#/components/schemas/Profile"
    CreateAdoptionRequest:
      type: object
      properties:
        pet_id:
          type: string
          format: uuid
        description:
          type: string
      required: [pet_id]
    UpdateAdoptionRequest:
      type: object
      properties:
        status:
          type: string
          enum: [available, pending, adopted]
        adopter_id:
          type: string
          format: uuid
        description:
          type: string
    PaginatedAdoptions:
      allOf:
        - $ref: "#/components/schemas/PaginatedResponse"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Adoption"

    # --- InstaPet ---
...